+++ src/api/fetch.ts
@@ -19,6 +19,9 @@ export async function extractCurrentUserFromPage(): Promise<
     const user = userLoginMetaTags[0].getAttribute("content");
 
     if (user != null && user !== "") {
+      browser.storage.local.set({ MOST_RECENT_GITHUB_USER: user });
+      console.log("Stored user in browser local extension storage");
+
       return user;
     } else {
       return undefined;