+++ CHANGELOG.md
@@ -12,11 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - You can now see what other people are voting and commenting on via the Guide by clicking on the extension's icon in your browser's nav bar. The Guide currently features:
   - Firehose: an unfiltered stream showing how others are voting and commenting
   - Top 10: a list of top and controversial projects throughout GitHub
+  - Inbox: a list of comments that mention your GitHub username
 - You can now share your opinions on a repository by adding a comment
 - You can now cancel your vote by clicking on the selected vote button again
 - Show profile pictures next to voter counts
-
-### Changed
 
 ### Fixed